refactor matcher
