+++ easycv/models/detection/dab_detr/dab_detr_head.py
@@ -5,10 +5,10 @@ import numpy as np
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
-from scipy.optimize import linear_sum_assignment
 
 from easycv.models.builder import HEADS, build_transformer
-from easycv.models.detection.utils import (accuracy, box_cxcywh_to_xyxy,
+from easycv.models.detection.utils import (HungarianMatcher, accuracy,
+                                           box_cxcywh_to_xyxy,
                                            box_xyxy_to_cxcywh,
                                            generalized_box_iou)
 from easycv.models.utils import get_world_size, is_dist_avail_and_initialized
@@ -36,7 +36,6 @@ class DABDETRHead(nn.Module):
                  num_classes,
                  embed_dims,
                  query_dim=4,
-                 focal_alpha=0.25,
                  iter_update=True,
                  num_select=300,
                  bbox_embed_diff_each_layer=False,
@@ -56,12 +55,11 @@ class DABDETRHead(nn.Module):
         super(DABDETRHead, self).__init__()
 
         self.matcher = HungarianMatcher(
-            cost_dict=cost_dict, focal_alpha=focal_alpha)
+            cost_dict=cost_dict, cost_class_type='focal_loss_cost')
         self.criterion = SetCriterion(
             num_classes,
             matcher=self.matcher,
             weight_dict=weight_dict,
-            focal_alpha=focal_alpha,
             losses=['labels', 'boxes', 'cardinality'])
         self.postprocess = PostProcess(num_select=num_select)
         self.transformer = build_transformer(transformer)
@@ -266,87 +264,6 @@ class MLP(nn.Module):
         return x
 
 
-class HungarianMatcher(nn.Module):
-    """This class computes an assignment between the targets and the predictions of the network
-    For efficiency reasons, the targets don't include the no_object. Because of this, in general,
-    there are more predictions than targets. In this case, we do a 1-to-1 matching of the best predictions,
-    while the others are un-matched (and thus treated as non-objects).
-    """
-
-    def __init__(self, cost_dict, focal_alpha):
-        """Creates the matcher
-        Params:
-            cost_class: This is the relative weight of the classification error in the matching cost
-            cost_bbox: This is the relative weight of the L1 error of the bounding box coordinates in the matching cost
-            cost_giou: This is the relative weight of the giou loss of the bounding box in the matching cost
-        """
-        super().__init__()
-        self.cost_class = cost_dict['cost_class']
-        self.cost_bbox = cost_dict['cost_bbox']
-        self.cost_giou = cost_dict['cost_giou']
-        self.focal_alpha = focal_alpha
-        assert self.cost_class != 0 or self.cost_bbox != 0 or self.cost_giou != 0, 'all costs cant be 0'
-
-    @torch.no_grad()
-    def forward(self, outputs, targets):
-        """ Performs the matching
-        Params:
-            outputs: This is a dict that contains at least these entries:
-                 "pred_logits": Tensor of dim [batch_size, num_queries, num_classes] with the classification logits
-                 "pred_boxes": Tensor of dim [batch_size, num_queries, 4] with the predicted box coordinates
-            targets: This is a list of targets (len(targets) = batch_size), where each target is a dict containing:
-                 "labels": Tensor of dim [num_target_boxes] (where num_target_boxes is the number of ground-truth
-                           objects in the target) containing the class labels
-                 "boxes": Tensor of dim [num_target_boxes, 4] containing the target box coordinates
-        Returns:
-            A list of size batch_size, containing tuples of (index_i, index_j) where:
-                - index_i is the indices of the selected predictions (in order)
-                - index_j is the indices of the corresponding selected targets (in order)
-            For each batch element, it holds:
-                len(index_i) = len(index_j) = min(num_queries, num_target_boxes)
-        """
-        bs, num_queries = outputs['pred_logits'].shape[:2]
-
-        # We flatten to compute the cost matrices in a batch
-        out_prob = outputs['pred_logits'].flatten(
-            0, 1).sigmoid()  # [batch_size * num_queries, num_classes]
-        out_bbox = outputs['pred_boxes'].flatten(
-            0, 1)  # [batch_size * num_queries, 4]
-
-        # Also concat the target labels and boxes
-        tgt_ids = torch.cat([v['labels'] for v in targets])
-        tgt_bbox = torch.cat([v['boxes'] for v in targets])
-
-        # Compute the classification cost.
-        alpha = self.focal_alpha
-        gamma = 2.0
-        neg_cost_class = (1 - alpha) * (out_prob**gamma)
-        neg_cost_class = neg_cost_class * (-(1 - out_prob + 1e-8).log())
-        pos_cost_class = alpha * ((1 - out_prob)**gamma)
-        pos_cost_class = pos_cost_class * (-(out_prob + 1e-8).log())
-        cost_class = pos_cost_class[:, tgt_ids] - neg_cost_class[:, tgt_ids]
-
-        # Compute the L1 cost between boxes
-        cost_bbox = torch.cdist(out_bbox, tgt_bbox, p=1)
-
-        # Compute the giou cost betwen boxes
-        # import ipdb; ipdb.set_trace()
-        cost_giou = -generalized_box_iou(
-            box_cxcywh_to_xyxy(out_bbox), box_cxcywh_to_xyxy(tgt_bbox))
-
-        # Final cost matrix
-        C = self.cost_bbox * cost_bbox + self.cost_class * cost_class + self.cost_giou * cost_giou
-        C = C.view(bs, num_queries, -1).cpu()
-
-        sizes = [len(v['boxes']) for v in targets]
-        indices = [
-            linear_sum_assignment(c[i])
-            for i, c in enumerate(C.split(sizes, -1))
-        ]
-        return [(torch.as_tensor(i, dtype=torch.int64),
-                 torch.as_tensor(j, dtype=torch.int64)) for i, j in indices]
-
-
 def sigmoid_focal_loss(inputs,
                        targets,
                        num_boxes,
@@ -387,21 +304,19 @@ class SetCriterion(nn.Module):
         2) we supervise each pair of matched ground-truth / prediction (supervise class and box)
     """
 
-    def __init__(self, num_classes, matcher, weight_dict, focal_alpha, losses):
+    def __init__(self, num_classes, matcher, weight_dict, losses):
         """ Create the criterion.
         Parameters:
             num_classes: number of object categories, omitting the special no-object category
             matcher: module able to compute a matching between targets and proposals
             weight_dict: dict containing as key the names of the losses and as values their relative weight.
             losses: list of all the losses to be applied. See get_loss for list of available losses.
-            focal_alpha: alpha in Focal Loss
         """
         super().__init__()
         self.num_classes = num_classes
         self.matcher = matcher
         self.weight_dict = weight_dict
         self.losses = losses
-        self.focal_alpha = focal_alpha
 
     def loss_labels(self, outputs, targets, indices, num_boxes, log=True):
         """Classification loss (Binary focal loss)
@@ -430,10 +345,7 @@ class SetCriterion(nn.Module):
 
         target_classes_onehot = target_classes_onehot[:, :, :-1]
         loss_ce = sigmoid_focal_loss(
-            src_logits,
-            target_classes_onehot,
-            num_boxes,
-            alpha=self.focal_alpha,
+            src_logits, target_classes_onehot, num_boxes, alpha=0.25,
             gamma=2) * src_logits.shape[1] * self.weight_dict['loss_ce']
         losses = {'loss_ce': loss_ce}
 
@@ -482,11 +394,6 @@ class SetCriterion(nn.Module):
                 box_cxcywh_to_xyxy(target_boxes)))
         losses['loss_giou'] = loss_giou.sum(
         ) / num_boxes * self.weight_dict['loss_giou']
-
-        # # calculate the x,y and h,w loss
-        # with torch.no_grad():
-        #     losses['loss_xy'] = loss_bbox[..., :2].sum() / num_boxes
-        #     losses['loss_hw'] = loss_bbox[..., 2:].sum() / num_boxes
 
         return losses
