+++ easycv/models/detection/utils/__init__.py
@@ -4,5 +4,6 @@ from .boxes import (bbox2result, bbox_overlaps, bboxes_iou, box_cxcywh_to_xyxy,
                     box_xyxy_to_cxcywh, distance2bbox, generalized_box_iou,
                     postprocess)
 from .generator import MlvlPointGenerator
+from .matcher import HungarianMatcher
 from .misc import (accuracy, filter_scores_and_topk, fp16_clamp, interpolate,
                    multi_apply, output_postprocess, select_single_mlvl)
